Adds scoring and game state.
Adds a scoring system so that a set number of points is granted for each
Apple and Vim sprite touched. Those sprites are removed upon contact.
Touching a Dell or VSCode sprite will result in a game over. This also
necessitates adding state to the game, with an initial title screen and
a game over screen that displays the current overall high score for the
session and offers an option to play again.

'q' is added as a method for quitting the game from the title or game
over screen rather than just closing the window.